wire new image and metadata
