+++ plan_refactor.md
@@ -29,3 +29,4 @@ src--
     - [] create MetaData class (inherit for different modalities)
 - [] Contour manipulation and mousevents need to be their own logic so can be reused by different windows
 - [] ResultsPlot could have it's own dataclass maybe?
+- [] big problem still with uninitialized variables in gui.py. images_display? image_displayed??
